Exception handling - delete

+++ src/main/java/com/woka/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.woka.course.entities.User;
 import com.woka.course.repositories.UserRepository;
+import com.woka.course.services.exceptions.DatabaseException;
 import com.woka.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -19,23 +21,31 @@ public class UserService {
 	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
-	public void delete (Long id) {
-		repository.deleteById(id);
+
+	public void delete(Long id) {
+		try {
+			if (!repository.existsById(id))
+				throw new ResourceNotFoundException(id);
+			repository.deleteById(id);
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
-		updateData(entity,obj);
+		updateData(entity, obj);
 		return repository.save(entity);
 	}
 
@@ -43,7 +53,7 @@ public class UserService {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
-		
+
 	}
 
 }
